Extract envoy classes into their own module
